Fix uintptr_t for 32-bit mode is an unsigned long.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@315 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/include/vadefs.h
@@ -28,7 +28,7 @@ extern "C" {
   typedef unsigned __int64 uintptr_t;
 #endif
 #else
-  typedef unsigned int uintptr_t;
+  typedef unsigned long uintptr_t;
 #endif
 #define _UINTPTR_T_DEFINED
 #endif
